+++ examples/js/app.js
@@ -5,12 +5,13 @@
 
     .config(['$routeProvider', function ($routeProvider) {
       $routeProvider
-        .when('/', {
-          templateUrl: 'views/home.html'
-        })
         .when('/basic-example', {
           controller: 'BasicExampleCtrl',
           templateUrl: 'views/basic-example.html'
+        })
+        .when('/cloning', {
+          controller: 'CloningCtrl',
+          templateUrl: 'views/cloning.html'
         })
         .when('/connected-trees', {
           controller: 'ConnectedTreesCtrl',
@@ -20,9 +21,12 @@
           controller: 'FilterNodesCtrl',
           templateUrl: 'views/filter-nodes.html'
         })
+        .when('/nodrop', {
+          controller: 'BasicExampleCtrl',
+          templateUrl: 'views/nodrop.html'
+        })
         .otherwise({
-          redirectTo: '/'
+          redirectTo: '/basic-example'
         });
-    }])
-
+    }]);
 })();