+++ protractor.conf.js
@@ -1,0 +1,19 @@
+// An example configuration file.
+exports.config = {
+  baseUrl: 'http://localhost:9000',
+
+  // Capabilities to be passed to the webdriver instance.
+  capabilities: {
+    'browserName': 'firefox'
+  },
+
+  // Spec patterns are relative to the current working directly when
+  // protractor is called.
+  specs: ['e2e/**/*.js'],
+
+  // Options to be passed to Jasmine-node.
+  jasmineNodeOpts: {
+    showColors: true,
+    defaultTimeoutInterval: 30000
+  }
+};